pyproject.toml
[project]
name = "projeto-sqlalchemy-wtf"
version = "0.1.0"
description = ""
authors = [
    {name = "Caio Araujo",email = "caio.dev.araujo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "flask-mail (>=0.10.0,<0.11.0)"
]

[tool.poetry]
packages = [{include = "projeto_sqlalchemy_wtf", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
